requirements.txt
mysql-connector-python
telepot
python-dotenv